Update lastEditTime everywhere

+++ frontend/src/components/dataExchange/GoogleDriveSynchronizer.tsx
@@ -4,8 +4,8 @@ import {useContext, useState} from "react";
 import ToastContext from "../../contexts/ToastContext";
 import {deserializeProfile, serializeProfile} from "../../services/ProfileSerializer";
 import ProfileContext from "../../contexts/ProfileContext";
-import {markProfileDirty} from "../../services/Persistence";
 import {mergeProfiles} from "../../services/Merger";
+import {markProfileDirty} from "../../services/Persistence";
 
 function GoogleDriveSynchronizer() {
 
@@ -14,7 +14,7 @@ function GoogleDriveSynchronizer() {
     const {
         hasDriveAccess, requestDriveAccess,
         isDriveLoaded, isDriveAuthorizing, error,
-        fetchFileList, fetchFile, uploadFile, updateFile, deleteFile
+        fetchFileList, fetchFile, uploadFile, deleteFile
     } = useGoogleDrive();
 
     const [existingFile, setExistingFile] = useState<gapi.client.drive.File | "no file">();
@@ -32,6 +32,7 @@ function GoogleDriveSynchronizer() {
         return "Error while authorizing: " + error;
     }
     
+    //TODO: store token in local storage?
     if (!hasDriveAccess){
         return (
             <CButton color="primary" onClick={requestDriveAccess}>Authorize with Google</CButton>
@@ -79,7 +80,6 @@ function GoogleDriveSynchronizer() {
             });
     }
 
-
     if (existingFile === "no file"){
         return (
             <CButton color="primary" onClick={createProfileInDrive}>Create new profile in drive</CButton>
@@ -87,38 +87,11 @@ function GoogleDriveSynchronizer() {
     }
 
     const deleteProfileFromDrive = () => {
-        deleteFile(existingFile)
-            .then(r =>  showToast("Profile deleted", 'success'))
-            .catch(error => showToast("Error deleting: " + error, 'danger'));
-    };
-
-    const uploadProfileToDrive = () => {
-        const file = new File([serializeProfile(profile)], "profile.json", {
-            type: "text/plain",
-        });
-
-        updateFile({id: existingFile.id ?? "", file, metadata})
-            .then(r => showToast("Profile is uploaded", 'success'))
-            .catch(error => showToast("error updating: " + error, 'danger'));
-    };
-
-    const getProfileFromDrive = () => {
-        fetchFile<'arraybuffer'>(existingFile) //for some reason, 'text' and 'json' do not work and still return arraybuffer
-            .then(r =>
-            {
-                const decoder = new TextDecoder('utf-8');
-                const text = decoder.decode(new Uint8Array(r.data as ArrayBuffer));
-                const deserializedProfile = deserializeProfile(text);
-                if (!deserializedProfile){
-                    showToast("Profile from cloud is invalid", 'danger')
-                }
-                else {
-                    setProfile(deserializedProfile)
-                    markProfileDirty(deserializedProfile)
-                    showToast("Profile is updated", 'success')
-                }
-            }            )
-            .catch(error => showToast("error fetching: " + error, 'danger'));
+        if (window.confirm("Are you sure you want to delete your profile from Drive?")) {
+            deleteFile(existingFile)
+                .then(r => showToast("Profile deleted", 'success'))
+                .catch(error => showToast("Error deleting: " + error, 'danger'));
+        }
     };
 
     const smartMerge = () => {
@@ -134,7 +107,9 @@ function GoogleDriveSynchronizer() {
                 }
                 else {
                     try {
-                        setProfile(mergeProfiles(profile, deserializedProfile))
+                        const mergedProfile = mergeProfiles(profile, deserializedProfile);
+                        setProfile(mergedProfile);
+                        markProfileDirty(mergedProfile, true);
                     }
                     catch (e){
                         if (e instanceof Error) {
@@ -152,10 +127,8 @@ function GoogleDriveSynchronizer() {
 
     return (
         <div className="d-flex flex-column gap-5">
-            <CButton color="primary" onClick={smartMerge}>Smart merge</CButton>
-            <CButton color="primary" onClick={getProfileFromDrive}>Get profile from drive</CButton>
-            <CButton color="primary" onClick={uploadProfileToDrive}>Upload profile to drive</CButton>
-            <CButton color="danger" onClick={deleteProfileFromDrive}>Delete profile from drive</CButton>
+            <CButton color="primary" onClick={smartMerge}>Synchronize data</CButton>
+            <CButton className="small" color="danger" onClick={deleteProfileFromDrive}>Delete profile from drive</CButton>
         </div>
     );
 }
